requirements.txt
# MCP y comunicación
mcp>=1.0.0
httpx>=0.25.0

# Validación y configuración
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0

# JWT
pyjwt>=2.8.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
